# benchmarks: print results in MB/s as well