This makes it easier to interpret the results.

go test -bench=. -test.run=none

before:

    BenchmarkEncryptThenMac-2         	  300000	      3958 ns/op
    BenchmarkAuthenticateThenDecrypt-2	  500000	      3939 ns/op

after:

    BenchmarkEncryptThenMac-2         	  300000	      3906 ns/op	  16.38 MB/s
    BenchmarkAuthenticateThenDecrypt-2	  500000	      3934 ns/op	  16.27 MB/s